+++ R/soda.R
@@ -1,36 +1,35 @@
 globalVariables(c(
   "Time", "Count", "Sensor", "Date", "Date_Time", # melb_walk_fast()
-  "Sensor_ID", "Longitude", "Latitude", "Location_Type", "Year_Installed",
-  "sensorloc", "sensorid", "longitude", "latitude", "loctype", "xdate"
+  "direction_1", "direction_2", "installation_date", "note", "sensor_description",
+  "sensor_df", "sensor_id", "status", "latitude", "longitude",
+  "date_time", "time", "total_of_directions"
 ))
 
-#' API using Socrata to Melbourne pedestrian data
+#' API using Socrata to Melbourne pedestrian data (per hour)
 #'
-#' Provides API using Socrata to Melbourne pedestrian data in a tidy data form.
+#' @param year An integer or a vector of integers. By default, it's the current
+#' year.
+#' @param sensor Sensor names. By default, it pulls all the sensors. Use [pull_sensor]
+#' to see the available sensors.
+#' @param na.rm Logical. `FALSE` is the default suggesting to include `NA` in
+#' the dataset. `TRUE` removes the `NA`s.
+#' @param app_token Characters giving the application token. A limited number of
+#' requests can be made without an app token (`NULL`), but they are subject
+#' to much lower throttling limits than request that do include one. Sign up
+#' for an app token [here](https://data.melbourne.vic.gov.au/profile/app_tokens).
 #'
-#' @param year An integer or a vector of integers. By default, it's the current 
-#'   year.
-#' @param sensor Sensor names. By default, it pulls all the sensors. Use [lookup_sensor]
-#'   to see the available sensors.
-#' @param tz Deprecated. For this dataset, it should only be "Australia/Melbourne".
-#' @param na.rm Logical. `FALSE` is the default suggesting to include `NA` in 
-#'   the dataset. `TRUE` removes the `NA`s.
-#' @param app_token Characters giving the application token. A limited number of 
-#'    requests can be made without an app token (`NULL`), but they are subject 
-#'    to much lower throttling limits than request that do include one. Sign up
-#'    for an app token [here](https://data.melbourne.vic.gov.au/profile/app_tokens).
-#'
-#' @details It provides API using [Socrata](https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/mxb8-wn4w), 
-#'   where counts are uploaded on a monthly basis. The up-to-date data would be
-#'   till the previous month. The data is sourced from [Melbourne Open Data Portal](https://data.melbourne.vic.gov.au/Transport-Movement/Pedestrian-volume-updated-monthly-/b2ak-trbp). Please 
-#'   refer to Melbourne Open Data Portal for more details about the dataset and 
-#'   its policy.
+#' @details It provides the API using [Socrata](https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/mxb8-wn4w),
+#' where counts are uploaded on a monthly basis. The up-to-date data would be
+#' till the previous month. The data is sourced from
+#' [Melbourne Open Data Portal](https://data.melbourne.vic.gov.au/Transport-Movement/Pedestrian-volume-updated-monthly-/b2ak-trbp).
+#' Please refer to Melbourne Open Data Portal for more details about the dataset and
+#' its policy.
 #' @return A tibble including these variables as follows:
-#'   * Sensor: Sensor name (46 sensors up to date)
-#'   * Date_Time: Date time when the pedestrian counts are recorded
-#'   * Date: Date associated with Date_Time
-#'   * Time: Time of day
-#'   * Count: Hourly counts
+#' * `Sensor`: Sensor name
+#' * `Date_Time`: Date time when the pedestrian counts are recorded
+#' * `Date`: Date associated with `date_Time`
+#' * `Time`: Time of day
+#' * `Count`: Hourly counts
 #'
 #' @export
 #' @seealso [melb_walk]
@@ -39,15 +38,12 @@ globalVariables(c(
 #' \dontrun{
 #' # Retrieve the year 2017
 #' melb_walk_fast(year = 2017)
-#'   
+#'
 #' # Retrieve the year 2017 for Southern Cross Station
 #' melb_walk_fast(year = 2017, sensor = "Southern Cross Station")
 #' }
-melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
+melb_walk_fast <- function(year = NULL, sensor = NULL, na.rm = FALSE,
   app_token = NULL) {
-  if (tz != "") {
-    warning("Argument `tz` ignored.")
-  }
   tz <- "Australia/Melbourne"
   this_year <- as.integer(format(Sys.Date(), "%Y"))
   if (is.null(year)) {
@@ -66,12 +62,12 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
   nsensors <- 50L
   if (!is.null(sensor)) {
     sensor_str <- paste(
-      vapply(sensor, function(x) paste0("'", x, "'"), character(1)), 
+      vapply(sensor, function(x) paste0("'", x, "'"), character(1)),
       collapse = ", "
     )
     query <- paste0(query, "AND sensor_name in", "(", sensor_str, ")")
     nsensors[] <- length(sensor) # overwrite nsensors
-  } 
+  }
   query <- paste0(query, " ORDER BY :id LIMIT 50000")
   limit <- 50000L
 
@@ -87,10 +83,10 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
       base_url <- paste0(base_url, app_token)
     }
     response <- httr::GET(base_url, query = list("$query" = update_query))
-    content <- httr::content(response, as = "text", type = "text/csv", 
+    content <- httr::content(response, as = "text", type = "text/csv",
       encoding = "UTF-8")
     dat <- dplyr::as_tibble(utils::read.csv(
-      textConnection(content), 
+      textConnection(content),
       colClasses = rep("character", 4L),
       stringsAsFactors = FALSE,
       nrows = limit
@@ -101,7 +97,7 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
 
   ped <- dplyr::bind_rows(lst_dat)
   ped <- dplyr::mutate(
-    ped, 
+    ped,
     Date_Time = as.POSIXct(strptime(Date_Time, format = "%Y-%m-%dT%H:%M:%S"),
       tz = tz)
   )
@@ -123,14 +119,14 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
       Sensor
     )
     ped <- dplyr::mutate(
-      ped, 
+      ped,
       Date = as.Date.POSIXct(Date_Time, tz = tz),
       Count = as.integer(Count),
       Time = as.integer(substr(Date_Time, 12, 13))
     )
   } else {
     ped <- dplyr::mutate(
-      ped, 
+      ped,
       Date = as.Date.POSIXct(Date_Time, tz = tz),
       Count = as.integer(Count),
       Time = as.integer(Time),
@@ -141,24 +137,75 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
   dplyr::arrange(ped, Date_Time)
 }
 
+#' API using Socrata to Melbourne pedestrian data with directions (per minute)
+#'
+#' @inheritParams melb_walk_fast
+#'
+#' @details It provides the API using [Socrata](https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/d6mv-s43h),
+#' to access minute by minute directional pedestrian counts for *the last hour*
+#' from pedestrian sensor devices located across the city. The data is updated
+#' every 15 minutes.
+#'
+#' Columns `sensor_id`, `direction_1`, and `direction_2` can be used to join
+#' the data with the Sensor Locations dataset which details the location, status,
+#' and directional readings of sensors, which can be obtained from [pull_sensor()].
+#'
+#' @return A tibble including these variables as follows:
+#' * `sensor_id`: Sensor name
+#' * `date_time`: Date time when the pedestrian counts are recorded
+#' * `date`: Date associated with `date_time`
+#' * `time`: Time of day
+#' * `direction_1`: Direction 1 sensor reading (count of pedestrians)
+#' * `direction_2`: Direction 2 sensor reading (count of pedestrians)
+#' * `total_of_directions`: Total sensor reading i.e. direction 1+2 (count of pedestrians)
+#'
+#' @seealso [pull_sensor()]
+#'
+#' @examples
+#' \dontrun{
+#' melb_walk_directional()
+#' }
+melb_walk_directional <- function(app_token = NULL) {
+  tz <- "Australia/Melbourne"
+  base_url <- "https://data.melbourne.vic.gov.au/resource/d6mv-s43h.csv"
+  sel_cols <- paste(
+    "SELECT sensor_id", "date_time", "date", "time", "direction_1",
+    "direction_2", "total_of_directions", sep = ", "
+  )
+
+  query <- paste0(sel_cols, " ORDER BY :id LIMIT 10000")
+  if (!is.null(app_token)) {
+    app_token <- paste0("$$app_token=", app_token)
+    base_url <- paste0(base_url, app_token)
+  }
+  response <- httr::GET(base_url, query = list("$query" = query))
+  content <- httr::content(response, as = "text", type = "text/csv",
+    encoding = "UTF-8")
+  ped <- dplyr::as_tibble(utils::read.csv(
+    textConnection(content),
+    colClasses = rep("character", 7L),
+    stringsAsFactors = FALSE,
+    nrows = 10000
+  ))
+  dplyr::mutate(
+    ped,
+    date_time = as.POSIXct(strptime(date_time, format = "%Y-%m-%dT%H:%M:%S"),
+      tz = tz),
+    date = as.Date.POSIXct(date_time, tz = tz),
+    time = hms::parse_hm(time),
+    direction_1 = as.integer(direction_1),
+    direction_2 = as.integer(direction_2),
+    total_of_directions = as.integer(total_of_directions)
+  )
+}
+
 #' API using Socrata to Melbourne pedestrian sensor locations
 #'
 #' Provides API using Socrata to Melbourne pedestrian sensor locations.
 #'
-#' @param app_token Characters giving the application token. A limited number of 
-#'    requests can be made without an app token (`NULL`), but they are subject 
-#'    to much lower throttling limits than request that do include one. Sign up
-#'    for an app token [here](https://data.melbourne.vic.gov.au/profile/app_tokens).
+#' @inheritParams melb_walk_fast
 #'
-#' @details It provides API using [Socrata](https://dev.socrata.com/foundry/data.melbourne.vic.gov.au/xbm5-bb4n).
-#'
-#' @return A data frame including these variables as follows:
-#'   * Sensor: Sensor name (43 sensors up to date)
-#'   * Sensor_ID: Sensor identifier
-#'   * Longitude: Longitude
-#'   * Latitude: Latitude
-#'   * Location_Type: Location type
-#'   * Year_Installed: Year installed
+#' @details It provides API using [Socrata](https://data.melbourne.vic.gov.au/resource/h57g-5234).
 #'
 #' @export
 #' @seealso [melb_walk_fast]
@@ -168,29 +215,29 @@ melb_walk_fast <- function(year = NULL, sensor = NULL, tz = "", na.rm = FALSE,
 #' pull_sensor()
 #' }
 pull_sensor <- function(app_token = NULL) {
-  base_url <- "https://data.melbourne.vic.gov.au/resource/xbm5-bb4n.csv"
+  base_url <- "https://data.melbourne.vic.gov.au/resource/h57g-5234.csv"
   p_url <- httr::parse_url(base_url)
   if (!is.null(app_token)) p_url$query$`$$app_token` <- app_token
   response <- httr::GET(p_url)
-  content <- httr::content(response, as = "text", type = "text/csv", 
+  content <- httr::content(response, as = "text", type = "text/csv",
     encoding = "UTF-8")
   sensor_info <- utils::read.csv(
-    textConnection(content), 
-    colClasses = rep("character", 10L),
+    textConnection(content),
+    colClasses = rep("character", 11L),
     stringsAsFactors = FALSE,
-    nrows = 50L # give a buffer to 43
+    na.strings = "",
+    nrows = 60L
   )
   sensor_info <- dplyr::select(
-    sensor_info, sensorloc, sensorid, longitude, latitude, loctype, xdate
+    sensor_info, sensor = sensor_description, sensor_id, longitude, latitude, direction_1, direction_2, installation_date, status, note
   )
-  colnames(sensor_info) <- c("Sensor", "Sensor_ID", "Longitude", "Latitude",
-    "Location_Type", "Year_Installed")
   sensor_info <- dplyr::mutate(
     sensor_info,
-    Sensor_ID = as.integer(Sensor_ID),
-    Longitude = as.numeric(Longitude),
-    Latitude = as.numeric(Latitude),
-    Year_Installed = as.integer(Year_Installed)
+    longitude = as.numeric(longitude),
+    latitude = as.numeric(latitude),
+    installation_date = as.POSIXct(
+      strptime(installation_date, format = "%Y-%m-%dT%H:%M:%S"),
+      tz = "Australia/Melbourne")
   )
   sensor_info
 }