+++ NEWS.md
@@ -1,3 +1,9 @@
+# rwalkr 0.5.0
+
+* New function `melb_walk_directional()` to access minute by minute directional pedestrian counts for the last hour from pedestrian sensor devices located across the city.
+* Fixed `pull_sensor()` due to the Socrata API URL change.
+* Defunct `lookup_sensor()`, `walk_melb()`, `run_melb()` and `shine_melb()` in favour of suffixed function names.
+
 # rwalkr 0.4.0
 
 * Fixed status "400" for `melb_walk_fast()` (previously `run_melb()`) due to Socrata API changes.